lambda/load/requirements.txt
boto3==1.35.64
botocore==1.35.64
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
distlib==0.3.9
filelock==3.16.1
greenlet==3.1.1
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
moto==5.0.2
numpy==2.1.3
pandas==2.2.3
pbr==6.1.0
platformdirs==4.3.6
psycopg2-binary
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
responses==0.25.3
s3transfer==0.10.3
six==1.16.0
SQLAlchemy==2.0.36
stevedore==5.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
Werkzeug==3.1.3
xmltodict==0.14.2
